+++ Projeto/src/poo2/AcessorioFinal.java
@@ -14,11 +14,11 @@ public class AcessorioFinal extends Acessorio{ //Classe inserida no final de cad
         return ".jpeg";
     }
     
-    public double[] getMultiplicador(){
-        double[] mult = new double[3];
-        mult[0] = 0.0;
-        mult[1] = 0.0;
-        mult[2] = 0.0;
+    public int[] getMultiplicador(){
+        int[] mult = new int[3];
+        mult[0] = 0;
+        mult[1] = 0;
+        mult[2] = 0;
         return mult;
     }
 }
